requirements.txt
moviepy==1.0.3
requests
boto3
pydantic
typing-extensions
edge-tts
google-generativeai
python-dotenv
